Update services.sh
1. force to be executed by bash because the syntax of "echo -e" does something else with some other shell
2. exit if no service is selected
3. return to service list if no action is selected

+++ services.sh
@@ -1,17 +1,26 @@
+#!/usr/bin/env bash
+
 selected=$(systemctl list-unit-files --no-pager --type=service --no-legend | dmenu | awk '{print $1;}')
 
 selected=$(echo $selected | awk '{print $1;}')
 
-action=$(echo -e "start\nstop\nrestart" | dmenu)
+if [ x$selected != x ] ; then
 
-case "$action" in
-    "start")
-        sudo systemctl start $selected
-        ;;
-    "stop")
-        sudo systemctl stop $selected
-        ;;
-    "restart")
-        sudo systemctl restart $selected
-        ;;
-esac
+    action=$(echo -e "start\nstop\nrestart" | dmenu)
+    
+    case "$action" in
+        "start")
+            sudo systemctl start $selected
+            ;;
+        "stop")
+            sudo systemctl stop $selected
+            ;;
+        "restart")
+            sudo systemctl restart $selected
+            ;;
+        "")
+            exec $0 "$@"
+            ;;
+    esac
+
+fi
